+++ packages/application-shell/src/components/theme-switcher/theme-switcher.tsx
@@ -1,0 +1,14 @@
+import { ThemeProvider } from '@commercetools-uikit/design-system';
+import { useFeatureToggle } from '@flopflip/react-broadcast';
+import { themesOverrides } from '@commercetools-frontend/application-components';
+import { UI_REDESIGN } from '../../feature-toggles';
+
+const ThemeSwitcher = () => {
+  const isNewThemeEnabled = useFeatureToggle(UI_REDESIGN);
+  const theme = isNewThemeEnabled ? 'test' : 'default';
+  return (
+    <ThemeProvider theme={theme} themeOverrides={themesOverrides[theme]} />
+  );
+};
+
+export default ThemeSwitcher;