+++ website-components-playground/src/components/modal-controller.tsx
@@ -1,10 +1,20 @@
-import { useState } from 'react';
-import PropTypes from 'prop-types';
+import { type ReactNode, useState } from 'react';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
 import SecondaryButton from '@commercetools-uikit/secondary-button';
+
+type TModalControllerFunctionOptions = {
+  isOpen: boolean;
+  setIsOpen: (nextValue: boolean) => void;
+};
+type TModalControllerProps = {
+  containerId: string;
+  title: string;
+  buttonLabel: string;
+  children: (options: TModalControllerFunctionOptions) => ReactNode;
+};
 
 const GridContainer = styled.div`
   display: grid;
@@ -17,7 +27,7 @@ const PortalContainer = styled.div`
   flex: 1;
 `;
 
-const ModalController = (props) => {
+const ModalController = (props: TModalControllerProps) => {
   const [isOpen, setIsOpen] = useState(true);
   return (
     <>
@@ -49,11 +59,5 @@ const ModalController = (props) => {
   );
 };
 ModalController.displayName = 'ModalController';
-ModalController.propTypes = {
-  containerId: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-  buttonLabel: PropTypes.string.isRequired,
-  children: PropTypes.func.isRequired,
-};
 
 export default ModalController;