+++ packages/application-shell-connectors/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @commercetools-frontend/application-shell-connectors
 
+## 21.19.0
+
+### Patch Changes
+
+- Updated dependencies [[`671436ba3`](https://github.com/commercetools/merchant-center-application-kit/commit/671436ba322ccbbed3301f54e73c0d669d026e0b)]:
+  - @commercetools-frontend/sentry@21.19.0
+  - @commercetools-frontend/constants@21.19.0
+
 ## 21.18.1
 
 ### Patch Changes