+++ packages/jest-preset-mc-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-frontend/jest-preset-mc-app
 
+## 21.19.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/babel-preset-mc-app@21.19.0
+
 ## 21.18.1
 
 ### Patch Changes