+++ packages/jest-stylelint-runner/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @commercetools-frontend/jest-stylelint-runner
 
+## 21.19.0
+
 ## 21.18.1
 
 ## 21.18.0