+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-frontend/sdk
 
+## 21.19.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/constants@21.19.0
+
 ## 21.18.1
 
 ### Patch Changes